+++ _10_OO_Demo/Product.cs
@@ -5,9 +5,9 @@ namespace OurStore;
 public class Product {
     public int ID { get; set; }
     public string Description { get; set; }
-    public double Price {get; set;}
+    public decimal Price {get; set;}
 
-    public Product (int id, string description, double price) {
+    public Product (int id, string description, decimal price) {
         ID = id;
         Description = description;
         Price = price;
